+++ docs/CHANGELOG.rst
@@ -38,6 +38,8 @@ Fixed
 -----
 - Fix ``Sample.get_reads()`` utility method
 - Fix ``duplicate`` endpoint now returns background task object
+- The data endpoint serializes ``collection`` information only on the top level, the
+  ``collection`` entry  inside ``entity`` is now a collection primary key
 
 
 ===================